requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.1.1
gunicorn==20.0.4
asgiref==3.4.1
Django==3.2.5
Pillow==8.3.0
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-decouple==3.4
urllib3==1.26.5
